Refactor composer dropdowns
